OPIK-1386 refactor EntityType out of FeedbackScoreDAO
