+++ apps/opik-backend/src/main/java/com/comet/opik/api/ExperimentSearchCriteria.java
@@ -1,14 +1,13 @@
 package com.comet.opik.api;
 
 import com.comet.opik.api.sorting.SortingField;
+import com.comet.opik.domain.EntityType;
 import lombok.Builder;
 import lombok.NonNull;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
-
-import static com.comet.opik.domain.FeedbackScoreDAO.EntityType;
 
 @Builder(toBuilder = true)
 public record ExperimentSearchCriteria(String name, UUID datasetId, @NonNull EntityType entityType,
